fix(context): fix multiple animations when jumpToNextStep() was called
In the last step,
every time the jumpToNextStep() function was called,
the animation from multistep was triggered,
causing problems with the user experience.

The solution was to check if the user was in the last step,
so that the animation would only be called once.
To do this, it was necessary to use another useEffect() in conjunction
with countSteps() and
check appropriately within jumpToNextStep()
whether the user is in the last step.